Web Client (Delete method fix)

+++ Spring-11-Rest-Jackson/src/main/java/com/cybertek/controller/WebFluxController.java
@@ -13,7 +13,12 @@ import reactor.core.publisher.Mono;
 
 @RestController
 public class WebFluxController {
-
+    /*
+        Spring WebFlux implements reactive programming via Reactive Streams.
+        In Spring WebFlux, the data returned from any operation is packed into a reactive stream.
+        There are two types that embody this approach and are the building blocks in WebFlux applications - Mono and Flux.
+     */
+    // Creating WebClient Instance
     private WebClient webClient = WebClient.builder().baseUrl("http://localhost:8080").build();
 
     private MovieCinemaRepository movieCinemaRepository;
@@ -24,11 +29,15 @@ public class WebFluxController {
         this.genreRepository = genreRepository;
     }
 
+    // Flux is a stream which returns zero or more items (0..N).
+    // Used when we're expecting multiple results or a collection of some sort.
     @GetMapping("/flux-movie-cinemas")
     public Flux<MovieCinema> readAllCinemaFlux() {
         return Flux.fromIterable(movieCinemaRepository.findAll());
     }
 
+    // Mono is a stream which returns zero items or a single item (0..1)
+    // Used when we're expecting a single (or none) result, such as retrieving a unique user from the database
     @GetMapping("/mono-movie-cinema/{id}")
     public Mono<MovieCinema> readByIdMono(@PathVariable("id") Long id) {
         return Mono.just(movieCinemaRepository.findByMovieCinemaId(id));
@@ -59,14 +68,21 @@ public class WebFluxController {
     }
 
     //------------------------- WEBCLIENT EXAMPLE -------------------------//
+    /*
+        WebClient is a reactive, non-blocking client for HTTP requests with a functional-style API client and Reactive Streams support.
+        By comparison to the RestTemplate, WebClient is:
+            1. Non-blocking, reactive and supports higher concurrency with less hardware resources
+            2. Provides a functional API that takes advantage of Java 8 lambdas
+            3. Supports both synchronous and asynchronous scenarios
+     */
 
     @GetMapping("/flux")
     public Flux<MovieCinema> readWithWebClient(){
         return webClient
-                .get()
-                .uri("/flux-movie-cinemas")
-                .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                .retrieve()
+                .get()                                                                  //
+                .uri("/flux-movie-cinemas")                                          // Setting request URI
+                .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)     // Setting request headers and authentication details
+                .retrieve()                                                             // retrieve()method directly performs the HTTP request and retrieve the response body.
                 .bodyToFlux(MovieCinema.class);
     }
 
@@ -96,13 +112,17 @@ public class WebFluxController {
         return webClient
                 .post()
                 .uri("/create-genre")
-                .body(Mono.just(genre), Genre.class)
+                .body(Mono.just(genre), Genre.class)                                       // Setting the request body
                 .retrieve()
                 .bodyToMono(Genre.class);
     }
 
     @DeleteMapping("/delete/{id}")
-    public Mono<Void> deleteGenreWebClient(@PathVariable("id") Long id) {
+    public Mono<Void> deleteGenreWebClient(@PathVariable("id") Long id) throws Exception {
+
+        Integer countGenres = genreRepository.countGenresNativeQuery(id);
+        if (countGenres > 0) { throw new RuntimeException("Genre can't be deleted, it is linked with movie(s)"); }
+
         return webClient
                 .delete()
                 .uri("/delete-genre/{id}", id)
